+++ azure-mgmt-containerregistry/azure/mgmt/containerregistry/v2018_09_01/models/file_task_step_update_parameters.py
@@ -17,6 +17,12 @@ class FileTaskStepUpdateParameters(TaskStepUpdateParameters):
 
     All required parameters must be populated in order to send to Azure.
 
+    :param context_path: The URL(absolute or relative) of the source context
+     for the task step.
+    :type context_path: str
+    :param context_access_token: The token (git PAT or SAS token of storage
+     account blob) associated with the context for a step.
+    :type context_access_token: str
     :param type: Required. Constant filled by server.
     :type type: str
     :param task_file_path: The task template/definition file path relative to
@@ -29,11 +35,6 @@ class FileTaskStepUpdateParameters(TaskStepUpdateParameters):
      when running a task.
     :type values:
      list[~azure.mgmt.containerregistry.v2018_09_01.models.SetValue]
-    :param context_path: The URL(absolute or relative) of the source context
-     for the build task.
-     If it is relative, the context will be relative to the source repository
-     URL of the build task.
-    :type context_path: str
     """
 
     _validation = {
@@ -41,11 +42,12 @@ class FileTaskStepUpdateParameters(TaskStepUpdateParameters):
     }
 
     _attribute_map = {
+        'context_path': {'key': 'contextPath', 'type': 'str'},
+        'context_access_token': {'key': 'contextAccessToken', 'type': 'str'},
         'type': {'key': 'type', 'type': 'str'},
         'task_file_path': {'key': 'taskFilePath', 'type': 'str'},
         'values_file_path': {'key': 'valuesFilePath', 'type': 'str'},
         'values': {'key': 'values', 'type': '[SetValue]'},
-        'context_path': {'key': 'contextPath', 'type': 'str'},
     }
 
     def __init__(self, **kwargs):
@@ -53,5 +55,4 @@ class FileTaskStepUpdateParameters(TaskStepUpdateParameters):
         self.task_file_path = kwargs.get('task_file_path', None)
         self.values_file_path = kwargs.get('values_file_path', None)
         self.values = kwargs.get('values', None)
-        self.context_path = kwargs.get('context_path', None)
         self.type = 'FileTask'