+++ azure-graphrbac/HISTORY.rst
@@ -3,6 +3,27 @@
 Release History
 ===============
 
+0.52.0 (2018-10-29)
++++++++++++++++++++
+
+**Bugfix**
+
+- Add missing required_resource_access in Application
+
+0.51.1 (2018-10-16)
++++++++++++++++++++
+
+**Bugfix**
+
+- Fix sdist broken in 0.50.0 and 0.51.0. No code change.
+
+0.51.0 (2018-10-11)
++++++++++++++++++++
+
+**Features**
+
+- Add delete group/application owner
+
 0.50.0 (2018-10-10)
 +++++++++++++++++++