+++ azure-mgmt-eventgrid/HISTORY.rst
@@ -3,6 +3,20 @@
 Release History
 ===============
 
+2.0.0rc2 (2018-10-24)
++++++++++++++++++++++
+
+**Features**
+
+- Model EventSubscriptionFilter has a new parameter advanced_filters
+- Model EventSubscriptionUpdateParameters has a new parameter expiration_time_utc
+- Model EventSubscription has a new parameter expiration_time_utc
+- Added operation EventSubscriptionsOperations.list_by_domain_topic
+- Added operation group DomainTopicsOperations
+- Added operation group DomainsOperations
+
+Internal API version is 2018-09-15-preview
+
 2.0.0rc1 (2018-05-04)
 +++++++++++++++++++++
 
@@ -14,9 +28,9 @@ Release History
 - delivering events to Azure Storage queue and Azure hybrid connections
 - deadlettering
 - retry policies
-- manual subscription validation handshake validation. 
+- manual subscription validation handshake validation.
 
-Internal API version is 2018-05-01-preview 
+Internal API version is 2018-05-01-preview
 
 1.0.0 (2018-04-26)
 ++++++++++++++++++
@@ -39,7 +53,7 @@ This version uses a next-generation code generator that *might* introduce breaki
 
   - Return type changes from `msrestazure.azure_operation.AzureOperationPoller` to `msrest.polling.LROPoller`. External API is the same.
   - Return type is now **always** a `msrest.polling.LROPoller`, regardless of the optional parameters used.
-  - The behavior has changed when using `raw=True`. Instead of returning the initial call result as `ClientRawResponse`, 
+  - The behavior has changed when using `raw=True`. Instead of returning the initial call result as `ClientRawResponse`,
     without polling, now this returns an LROPoller. After polling, the final resource will be returned as a `ClientRawResponse`.
   - New `polling` parameter. The default behavior is `Polling=True` which will poll using ARM algorithm. When `Polling=False`,
     the response of the initial call will be returned without polling.